requirements.txt
Flask~=2.2.2
pip~=21.3.1
wheel~=0.37.1
Jinja2~=3.1.2
setuptools~=60.2.0
kafka-python~=2.0.2
requests~=2.28.1